refactor(assay): extract functions.rs module
